% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/fastcmh.R
\name{DemoFastCMH}
\alias{DemoFastCMH}
\title{Demo of FastCMH}
\usage{
DemoFastCMH(folder = "fastcmhdemo/")
}
\arguments{
\item{folder}{The folder in which the data for the demo will be saved. Default is the current directory, "./". The demo data will created in folder/data and the results will be saved in folder/results as an RData file}
}
\description{
This function runs a demo for FastCMH, by first creating a sample data set and then running FastCMH on this data set.
}
\section{Details}{

This function will first create a sample data set in the folder/data, and will then run FastCMH on this data set, before saving the results in folder/results. The method runs in several steps, with each step showing the R code that can be used to do the step, then running that R code, and then waiting for the user to press enter before moving onto the next step.
}
\examples{
DemoFastCMH("../fastcmhdemodata")
DemoFastCMH() #default folder is ./fastcmhdemo
}

